add commands

+++ src/main/java/com/chcodes/demo/repo/CommandRepo.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.chcodes.demo.domain.Command;
 
-public interface CommandRepo extends JpaRepository<Command, Integer> {
+public interface CommandRepo extends JpaRepository<Command, Long> {
 
 	@Query("select c from Command c where c.user.id=:x")
 	List<Command> findByUser(@Param("x") int id);
